refactor(cli): extract web-UI build and PE-integrity helpers from hermes_cli/main.py (shard s3)
