+++ hermes_cli/main.py
@@ -5182,120 +5182,6 @@ def _sweep_stale_bytecode_if_checkout_changed() -> None:
         logger.debug("Stale-bytecode launch sweep failed: %s", exc)
 
 
-def _web_ui_build_needed(web_dir: Path) -> bool:
-    """Return True if the web UI dist is missing or its source content changed.
-
-    Uses a SHA-256 content hash of the web source tree (the same approach
-    ``_desktop_build_needed()`` already uses for the Electron build), NOT
-    mtime comparison. ``git checkout`` / ``git pull`` / ``hermes update``
-    rewrite source mtimes without changing content, which made the old
-    mtime check unreliable in both directions: it could skip a rebuild when
-    source had genuinely changed (serving a stale dashboard) and force a
-    rebuild when nothing had. A content hash is stable across mtime churn.
-
-    The dashboard source lives under ``web/`` but Vite outputs to
-    ``hermes_cli/web_dist/`` (per vite.config.ts outDir), NOT ``web/dist/``,
-    so the dist directory is never part of the hashed source tree.
-    """
-    project_root = web_dir.parent.parent if web_dir.parent.name == "apps" else web_dir.parent
-    dist_dir = project_root / "hermes_cli" / "web_dist"
-    sentinel = dist_dir / ".vite" / "manifest.json"
-    if not sentinel.exists():
-        sentinel = dist_dir / "index.html"
-    if not sentinel.exists():
-        return True
-    stamp_file = _web_ui_stamp_path()
-    if not stamp_file.is_file():
-        return True
-    try:
-        stamp_data = json.loads(stamp_file.read_text(encoding="utf-8"))
-    except (OSError, json.JSONDecodeError):
-        return True
-    if not isinstance(stamp_data, dict):
-        return True
-    saved_hash = stamp_data.get("contentHash")
-    if not saved_hash:
-        return True
-    return _compute_web_ui_content_hash(project_root, web_dir) != saved_hash
-
-
-def _compute_web_ui_content_hash(project_root: Path, web_dir: Path) -> str:
-    """Return a SHA-256 hex digest of the web UI source tree.
-
-    Covers ``web_dir`` (the dashboard frontend source) plus the root
-    ``package.json`` / ``package-lock.json`` (workspace config that
-    determines dependency resolution). Mirrors
-    ``_compute_desktop_content_hash()``: ignored paths (``node_modules/``,
-    ``dist/``, ``*.pyc``, ...) are skipped via the repo-root ``.gitignore``
-    so build output never feeds back into its own staleness check.
-    """
-    h = hashlib.sha256()
-
-    def _hash_file(path: Path) -> None:
-        rel = str(path.relative_to(project_root))
-        h.update(rel.encode())
-        h.update(b"\0")
-        try:
-            with open(path, "rb") as f:
-                for chunk in iter(lambda: f.read(65536), b""):
-                    h.update(chunk)
-        except OSError:
-            pass
-        h.update(b"\0")
-
-    from pathspec import PathSpec
-
-    gitignore = project_root / ".gitignore"
-    lines: list[str] = []
-    if gitignore.is_file():
-        lines = gitignore.read_text(encoding="utf-8").splitlines()
-    spec = PathSpec.from_lines("gitignore", lines)
-
-    # Root workspace config (single package-lock.json covers all workspaces).
-    for name in ("package.json", "package-lock.json"):
-        p = project_root / name
-        if p.is_file():
-            rel = str(p.relative_to(project_root))
-            if not spec.match_file(rel):
-                _hash_file(p)
-
-    # Walk the web source tree, pruning ignored directories in-place so we
-    # never descend into node_modules/ or a stray dist/. Sort filenames for
-    # a deterministic, order-independent digest.
-    for dirpath, dirnames, filenames in os.walk(web_dir, topdown=True):
-        dirnames[:] = [
-            d for d in dirnames
-            if not spec.match_file(str((Path(dirpath) / d).relative_to(project_root)))
-        ]
-        for fn in sorted(filenames):
-            fp = Path(dirpath) / fn
-            rel = str(fp.relative_to(project_root))
-            if not spec.match_file(rel):
-                _hash_file(fp)
-
-    return h.hexdigest()
-
-
-def _web_ui_stamp_path() -> Path:
-    """Return the path to the web UI build stamp file under $HERMES_HOME."""
-    from hermes_constants import get_hermes_home
-    return get_hermes_home() / "web-ui-build-stamp.json"
-
-
-def _write_web_ui_build_stamp(project_root: Path, web_dir: Path) -> None:
-    """Write the web UI build stamp after a successful build."""
-    stamp_file = _web_ui_stamp_path()
-    try:
-        stamp_file.parent.mkdir(parents=True, exist_ok=True)
-        from datetime import datetime, timezone
-        stamp_data = {
-            "contentHash": _compute_web_ui_content_hash(project_root, web_dir),
-            "builtAt": datetime.now(timezone.utc).isoformat(),
-        }
-        stamp_file.write_text(json.dumps(stamp_data, indent=2) + "\n", encoding="utf-8")
-    except Exception as exc:
-        # Never let stamp-writing block or fail a build.
-        logger.debug("Failed to write web UI build stamp: %s", exc)
 
 
 def _run_with_idle_timeout(
@@ -5581,356 +5467,6 @@ def _run_npm_watching_for_engine_failure(
     return subprocess.CompletedProcess(cmd, returncode, None, "".join(captured))
 
 
-def _missing_web_build_tool(output: str) -> str | None:
-    """Return the build tool a failed ``npm run build`` could not resolve.
-
-    Each shell words this differently: ``sh: 1: tsc: not found`` (dash),
-    ``vite: command not found`` (bash/zsh), and ``'tsc' is not recognized as
-    an internal or external command`` (cmd.exe).
-    """
-    lowered = output.lower()
-    for tool in ("tsc", "vite"):
-        if any(
-            phrase in lowered
-            for phrase in (
-                f"{tool}: not found",
-                f"{tool}: command not found",
-                f"'{tool}' is not recognized",
-            )
-        ):
-            return tool
-    return None
-
-
-def _build_web_ui(web_dir: Path, *, fatal: bool = False) -> bool:
-    """Build the web UI frontend if npm is available, serializing across processes.
-
-    Concurrent dashboard boots (e.g. the desktop app's retry loop after a
-    readiness timeout) used to each spawn their own ``npm install`` +
-    ``vite build`` over the same tree; the parallel builds starved each
-    other, none finished, the dist sentinel never advanced, and every new
-    boot re-triggered the build. One process builds under an exclusive
-    flock; the rest serve the existing dist (stale is acceptable) or, when
-    no dist exists yet, block until the builder finishes.
-
-    Staleness is checked once, inside :func:`_do_build_web_ui`, after the
-    lock is held — so a process that queued behind the builder skips the
-    rebuild, and the (os.walk-based) check runs at most once per boot.
-    """
-    if not (web_dir / "package.json").exists():
-        return True
-    try:
-        import fcntl
-    except ImportError:
-        # Windows: no flock — fall through to the unserialized build.
-        return _do_build_web_ui(web_dir, fatal=fatal)
-    project_root = web_dir.parent.parent if web_dir.parent.name == "apps" else web_dir.parent
-    dist_index = project_root / "hermes_cli" / "web_dist" / "index.html"
-    try:
-        lock_file = open(project_root / ".web_ui_build.lock", "a", encoding="utf-8")
-    except OSError:
-        return _do_build_web_ui(web_dir, fatal=fatal)
-    try:
-        try:
-            fcntl.flock(lock_file.fileno(), fcntl.LOCK_EX | fcntl.LOCK_NB)
-        except OSError:
-            if dist_index.exists():
-                # Another process is already building — serve the current
-                # dist instead of piling a second build onto the same tree.
-                return True
-            # No dist at all (first-ever build): wait for the builder.
-            fcntl.flock(lock_file.fileno(), fcntl.LOCK_EX)
-        return _do_build_web_ui(web_dir, fatal=fatal)
-    finally:
-        lock_file.close()
-
-
-def _do_build_web_ui(web_dir: Path, *, fatal: bool = False) -> bool:
-    """Build the web UI frontend if npm is available.
-
-    Args:
-        web_dir: Path to the dashboard frontend source directory.
-        fatal: If True, print error guidance and return False on failure
-               instead of a soft warning (used by ``hermes web``).
-
-    Returns True if the build succeeded or was skipped (no package.json).
-    """
-    if not (web_dir / "package.json").exists():
-        return True
-
-    if not _web_ui_build_needed(web_dir):
-        return True
-
-    # Console-encoding-safe print: Windows consoles default to cp1252
-    # (or similar) and will raise UnicodeEncodeError on arrow / check
-    # glyphs unless PYTHONIOENCODING=utf-8 is set. Routing every print
-    # in this function through _say() with errors="replace" keeps the
-    # build path usable on a stock `py -m hermes_cli.main web` invocation.
-    def _say(text: str) -> None:
-        try:
-            print(text)
-        except UnicodeEncodeError:
-            encoding = getattr(sys.stdout, "encoding", None) or "ascii"
-            print(text.encode(encoding, errors="replace").decode(encoding, errors="replace"))
-
-    from hermes_constants import with_hermes_node_path
-
-    npm = _resolve_node_runtime_npm()
-    if not npm:
-        if fatal:
-            _say("Web UI frontend not built and npm is not available.")
-            _say("Install Node.js, then run:  cd web && npm install && npm run build")
-        return not fatal
-    build_env = with_hermes_node_path()
-    _say("→ Building web UI...")
-
-    def _relay(result: "subprocess.CompletedProcess") -> None:
-        """Print captured npm output so users can see *why* a step failed.
-
-        Windows users hitting `rm -rf` / `cp -r` errors (or any other
-        sync-assets / Vite failure) would otherwise see only ``Web UI
-        build failed`` with no hint of the underlying cause, because
-        the npm calls run with ``capture_output=True``.
-        """
-        for blob in (result.stdout, result.stderr):
-            if not blob:
-                continue
-            text = blob.decode("utf-8", errors="replace").rstrip() if isinstance(blob, bytes) else blob.rstrip()
-            if text:
-                _say(text)
-
-    npm_cwd = _workspace_root(web_dir)
-    # Scope the install to the web workspace only so that the full workspace
-    # graph (including apps/desktop with its Electron + node-pty deps) is never
-    # resolved here.  Without --workspace the root package.json's apps/* glob
-    # would pull in desktop on every web build. See #38772.
-    # When web/ has its own package-lock.json, _workspace_root() returns
-    # web_dir itself and --workspace would fail.  See #42973.
-    npm_workspace_args: tuple[str, ...] = () if npm_cwd == web_dir else ("--workspace", "web")
-    if _is_termux_startup_environment():
-        npm_cwd, npm_workspace_args = _termux_workspace_install_context(web_dir)
-
-    def _install_web_deps(*, silent: bool) -> "subprocess.CompletedProcess":
-        return _run_npm_install_deterministic(
-            npm,
-            npm_cwd,
-            extra_args=(*npm_workspace_args, "--silent", "--prefer-offline") if silent else (*npm_workspace_args, "--prefer-offline"),
-            env=build_env,
-        )
-
-    r1 = _install_web_deps(silent=True)
-    if r1.returncode != 0:
-        _say(
-            f"  {'✗' if fatal else '⚠'} Web UI npm install failed"
-            + ("" if fatal else " (hermes web will not be available)")
-        )
-        _relay(r1)
-        if fatal:
-            _say("  Run manually:  npm install --workspace web && npm run build -w web")
-        return False
-    # First attempt — stream output via idle-timeout helper (issue #33788).
-    # capture_output=True on a long Vite build looks identical to a hang;
-    # users react by rebooting, which leaves the editable install in a
-    # half-state. Streaming + idle-kill makes failures observable AND
-    # recoverable (the stale-dist fallback below handles the kill path).
-    r2 = _run_with_idle_timeout([npm, "run", "build"], cwd=web_dir, env=build_env)
-    if r2.returncode != 0:
-        # The install above can exit 0 while leaving the tree without a build
-        # toolchain — a lockfile-hash skip over a half-installed tree, or an
-        # interrupted link step. The generic retry below just reruns the same
-        # command, so `tsc: not found` survives it and the stale dist is
-        # served forever. Reinstall (non-silent, so the user sees it) first.
-        missing_tool = _missing_web_build_tool((r2.stdout or "") + (r2.stderr or ""))
-        if missing_tool:
-            _say(f"  ⚠ Build could not resolve {missing_tool} — reinstalling web dependencies...")
-            _install_web_deps(silent=False)
-            r2 = _run_with_idle_timeout([npm, "run", "build"], cwd=web_dir, env=build_env)
-        if r2.returncode != 0:
-            # Retry once after a short delay — covers boot-time races on Windows
-            # (antivirus scanning Node.js binaries, npm cache not ready, transient
-            # I/O when launched via Scheduled Task at logon). See issue #23817.
-            _time.sleep(3)
-            r2 = _run_with_idle_timeout([npm, "run", "build"], cwd=web_dir, env=build_env)
-
-    if r2.returncode != 0:
-        # _run_with_idle_timeout merges stderr into stdout; older callers
-        # using subprocess.run kept them split. Pull from whichever has
-        # content so the error surfaces regardless of which path produced
-        # the CompletedProcess.
-        build_output = (r2.stderr or "") + (r2.stdout or "")
-        stderr_preview = build_output.strip()
-        stderr_tail = "\n  ".join(stderr_preview.splitlines()[-10:]) if stderr_preview else ""
-        project_root = web_dir.parent.parent if web_dir.parent.name == "apps" else web_dir.parent
-        dist_dir = project_root / "hermes_cli" / "web_dist"
-        dist_index = dist_dir / "index.html"
-
-        # If a stale dist exists, serve it as a fallback instead of failing.
-        # A stale UI is far better than no UI for non-interactive callers
-        # (Windows Scheduled Tasks, CI) — issue #23817.
-        if dist_index.exists():
-            _say("  ⚠ Web UI build failed — serving stale dist as fallback")
-            if stderr_tail:
-                _say(f"  Build error:\n  {stderr_tail}")
-            return True
-
-        _say(
-            f"  {'✗' if fatal else '⚠'} Web UI build failed"
-            + ("" if fatal else " (hermes web will not be available)")
-        )
-        _relay(r2)
-        if fatal:
-            _say("  Run manually:  npm install --workspace web && npm run build -w web")
-        return False
-    _say("  ✓ Web UI built")
-    project_root = web_dir.parent.parent if web_dir.parent.name == "apps" else web_dir.parent
-    _write_web_ui_build_stamp(project_root, web_dir)
-    return True
-
-
-def _desktop_dist_exists(desktop_dir: Path) -> bool:
-    """Return True when a local desktop renderer build is present."""
-    return (desktop_dir / "dist" / "index.html").exists()
-
-
-# ---------------------------------------------------------------------------
-# Desktop build stamp — content-hash based skip logic
-# ---------------------------------------------------------------------------
-# The desktop Electron build is expensive.
-# Unlike the web UI (which uses mtime comparison), the desktop uses a
-# SHA-256 content hash of the source tree so that:
-#   - ``git checkout`` / ``git pull`` that touch mtimes but not content
-#     don't trigger a rebuild
-#   - ``hermes update`` can unconditionally call ``hermes desktop --build-only``
-#     and it will skip if nothing actually changed
-#   - ``hermes desktop`` (interactive launch) skips the build when the
-#     stamp matches, making repeated launches fast
-#
-# Stamp file: $HERMES_HOME/desktop-build-stamp.json
-# Schema:
-#   {
-#     "contentHash": "<sha256 hex of source files>",
-#     "sourceMode": true | false,
-#     "builtAt": "<ISO 8601>"
-#   }
-
-def _compute_desktop_content_hash(project_root: Path) -> str:
-    """Return a SHA-256 hex digest of all source files that feed the desktop build.
-
-    Covers ``apps/desktop/`` (excluding anything matched by .gitignore)
-    plus the root ``package.json`` / ``package-lock.json`` (workspace config
-    that determines dependency resolution for the desktop workspace).
-
-    Parses the repo-root ``.gitignore`` via *pathspec* so we automatically
-    skip ``node_modules/``, ``dist/``, ``*.pyc``, etc. without maintaining
-    a hardcoded skip-list.
-    """
-    h = hashlib.sha256()
-
-    def _hash_file(path: Path) -> None:
-        rel = str(path.relative_to(project_root))
-        h.update(rel.encode())
-        h.update(b"\0")
-        try:
-            with open(path, "rb") as f:
-                for chunk in iter(lambda: f.read(65536), b""):
-                    h.update(chunk)
-        except (OSError, IOError):
-            pass
-        h.update(b"\0")
-
-
-    from pathspec import PathSpec
-
-    gitignore = project_root / ".gitignore"
-    lines: list[str] = []
-    if gitignore.is_file():
-        lines = gitignore.read_text(encoding="utf-8").splitlines()
-    spec = PathSpec.from_lines("gitignore", lines)
-
-    # Root workspace config
-    for name in ("package.json", "package-lock.json"):
-        p = project_root / name
-        if p.is_file():
-            rel = str(p.relative_to(project_root))
-            if not spec.match_file(rel):
-                _hash_file(p)
-
-    # Walk apps/desktop/ — prune ignored directories in-place
-    desktop_dir = project_root / "apps" / "desktop"
-    for dirpath, dirnames, filenames in os.walk(desktop_dir, topdown=True):
-        # Prune ignored directories so we never descend into them
-        dirnames[:] = [
-            d for d in dirnames
-            if not spec.match_file(str((Path(dirpath) / d).relative_to(project_root)))
-        ]
-
-        for fn in sorted(filenames):
-            fp = Path(dirpath) / fn
-            rel = str(fp.relative_to(project_root))
-            if not spec.match_file(rel):
-                _hash_file(fp)
-
-    return h.hexdigest()
-
-
-def _desktop_stamp_path() -> Path:
-    """Return the path to the desktop build stamp file under $HERMES_HOME."""
-    from hermes_constants import get_hermes_home
-    return get_hermes_home() / "desktop-build-stamp.json"
-
-
-def _desktop_build_needed(desktop_dir: Path, project_root: Path, *, source_mode: bool) -> bool:
-    """Return True when the desktop build output is stale or missing.
-
-    Compares the current content hash against the saved stamp. Also returns
-    True if the expected build artifact doesn't exist (e.g. first run after
-    ``hermes update`` that pulled new source but hasn't built yet).
-    """
-    # If there's no build output at all, we definitely need to build
-    if source_mode:
-        if not _desktop_dist_exists(desktop_dir):
-            return True
-    else:
-        if _desktop_packaged_executable(desktop_dir) is None:
-            return True
-
-    stamp_file = _desktop_stamp_path()
-    if not stamp_file.is_file():
-        return True
-
-    try:
-        stamp_data = json.loads(stamp_file.read_text(encoding="utf-8"))
-    except (OSError, json.JSONDecodeError, KeyError):
-        return True
-
-    # If the mode changed (source vs packaged), force a rebuild
-    if stamp_data.get("sourceMode") != source_mode:
-        return True
-
-    saved_hash = stamp_data.get("contentHash")
-    if not saved_hash:
-        return True
-
-    current_hash = _compute_desktop_content_hash(project_root)
-    return current_hash != saved_hash
-
-
-def _write_desktop_build_stamp(project_root: Path, *, source_mode: bool) -> None:
-    """Write the desktop build stamp after a successful build."""
-    stamp_file = _desktop_stamp_path()
-    try:
-        stamp_file.parent.mkdir(parents=True, exist_ok=True)
-        content_hash = _compute_desktop_content_hash(project_root)
-        from datetime import datetime, timezone
-        stamp_data = {
-            "contentHash": content_hash,
-            "sourceMode": source_mode,
-            "builtAt": datetime.now(timezone.utc).isoformat(),
-        }
-        stamp_file.write_text(json.dumps(stamp_data, indent=2) + "\n", encoding="utf-8")
-    except Exception as exc:
-        # Never let stamp-writing block or fail a build
-        logger.debug("Failed to write desktop build stamp: %s", exc)
 
 
 def _desktop_packaged_executable(desktop_dir: Path) -> Optional[Path]:
@@ -5969,340 +5505,6 @@ def _desktop_packaged_executable(desktop_dir: Path) -> Optional[Path]:
     return max(existing, key=lambda p: p.stat().st_mtime)
 
 
-# ─── Desktop exe integrity gate (#69179) ────────────────────────────────────
-#
-# The desktop self-update chain (Desktop → hermes-setup --update →
-# `hermes update` → `hermes desktop --build-only` → relaunch) rebuilds
-# Hermes.exe on the end user's machine and used to verify only that the file
-# EXISTS before declaring success. A corrupt cached Electron zip whose
-# extraction produced a truncated electron.exe, an interrupted rcedit resource
-# rewrite, a disk-full pack, or a wrong-arch unpacked tree therefore shipped a
-# broken binary that Windows refuses to load ("This app can't run on your
-# computer" / 此应用无法在你的电脑上运行). These helpers parse the PE header —
-# no signature infrastructure required — so a structurally broken or
-# wrong-architecture Hermes.exe is caught BEFORE the updater replaces the
-# working app, and the previous build can be restored from the .bak tree that
-# apps/desktop/scripts/before-pack.mjs now preserves.
-
-_PE_MACHINE_I386 = 0x014C
-_PE_MACHINE_AMD64 = 0x8664
-_PE_MACHINE_ARM64 = 0xAA64
-
-_PE_MACHINE_NAMES = {
-    _PE_MACHINE_I386: "x86 (32-bit)",
-    _PE_MACHINE_AMD64: "x64 (AMD64)",
-    _PE_MACHINE_ARM64: "ARM64",
-}
-
-_PE_MACHINE_TO_NAME = {
-    _PE_MACHINE_ARM64: "ARM64",
-    _PE_MACHINE_AMD64: "AMD64",
-    _PE_MACHINE_I386: "X86",
-}
-
-# MACHINE_ATTRIBUTES bits (processthreadsapi.h). UserEnabled means the host
-# can run user-mode code of that machine type — natively or under emulation.
-_MACHINE_ATTRIBUTE_USER_ENABLED = 0x00000001
-
-
-def _windows_native_machine_from_iswow64() -> Optional[str]:
-    """Ask IsWow64Process2 for the OS-native machine (None if unavailable/fail).
-
-    ctypes defaults ``GetCurrentProcess``'s restype to ``c_int``, so the
-    current-process pseudo-handle ``(HANDLE)-1`` is truncated to
-    ``0xFFFFFFFF`` and zero-extended into a 64-bit invalid handle. On Win64
-    that makes ``IsWow64Process2`` fail with ``ERROR_INVALID_HANDLE`` (6),
-    which is exactly the residual Windows-on-ARM failure after #71218: the
-    gate fell through to ``PROCESSOR_ARCHITECTURE=AMD64`` (the emulated
-    process arch) and rejected a correctly-built ARM64 ``Hermes.exe``.
-    Binding ``restype``/``argtypes`` to ``wintypes.HANDLE`` keeps the full
-    ``0xFFFFFFFFFFFFFFFF`` pseudo-handle.
-    """
-    import ctypes
-    from ctypes import wintypes
-
-    kernel32 = ctypes.WinDLL("kernel32", use_last_error=True)
-    kernel32.GetCurrentProcess.restype = wintypes.HANDLE
-    kernel32.GetCurrentProcess.argtypes = []
-    kernel32.IsWow64Process2.argtypes = [
-        wintypes.HANDLE,
-        ctypes.POINTER(wintypes.USHORT),
-        ctypes.POINTER(wintypes.USHORT),
-    ]
-    kernel32.IsWow64Process2.restype = wintypes.BOOL
-
-    process_machine = wintypes.USHORT(0)
-    native_machine = wintypes.USHORT(0)
-    if not kernel32.IsWow64Process2(
-        kernel32.GetCurrentProcess(),
-        ctypes.byref(process_machine),
-        ctypes.byref(native_machine),
-    ):
-        return None
-    return _PE_MACHINE_TO_NAME.get(native_machine.value)
-
-
-def _windows_user_runnable_pe_machines() -> Optional[set]:
-    """PE machines this host can run in user mode, via GetMachineTypeAttributes.
-
-    This asks the question the integrity gate actually cares about — "can this
-    Windows host load a PE of machine X?" — instead of inferring it from a
-    host-architecture name. It is also the only documented API that reports
-    AMD64-on-ARM64 emulation support; ``IsWow64GuestMachineSupported`` only
-    answers for 32-bit guests.
-
-    Returns None when the API is unavailable (pre-Windows-11 build 22000) or
-    reports nothing runnable, so callers fall back to name-based detection.
-    """
-    import ctypes
-    from ctypes import wintypes
-
-    kernel32 = ctypes.WinDLL("kernel32", use_last_error=True)
-    kernel32.GetMachineTypeAttributes.argtypes = [
-        wintypes.USHORT,
-        ctypes.POINTER(ctypes.c_int),
-    ]
-    kernel32.GetMachineTypeAttributes.restype = ctypes.c_long
-
-    runnable = set()
-    for machine in (_PE_MACHINE_ARM64, _PE_MACHINE_AMD64, _PE_MACHINE_I386):
-        attributes = ctypes.c_int(0)
-        # HRESULT: zero is success, any nonzero value is a failure.
-        if kernel32.GetMachineTypeAttributes(machine, ctypes.byref(attributes)):
-            continue
-        if attributes.value & _MACHINE_ATTRIBUTE_USER_ENABLED:
-            runnable.add(machine)
-    return runnable or None
-
-
-def _windows_native_machine() -> str:
-    """The Windows host OS's NATIVE machine architecture, normalized upper.
-
-    ``platform.machine()`` reports the PROCESS architecture, which lies under
-    emulation: the desktop update chain runs an x64 hermes-setup.exe (and thus
-    x64 Python) on Windows-on-ARM devices, where ``platform.machine()``
-    returns ``AMD64`` even though the OS is ARM64. The #71119 integrity gate
-    then rejected the CORRECT ARM64 rebuild as an "architecture mismatch"
-    (#69179 follow-up report). Probe order:
-
-    1. ``IsWow64Process2`` with a correctly-typed current-process HANDLE
-       (#71218 + HANDLE-truncation fix). This is the only API that tells the
-       truth from an x64 process emulated on ARM64.
-    2. ``PROCESSOR_ARCHITEW6432`` / ``PROCESSOR_ARCHITECTURE`` — WOW64
-       (32-bit) hosts and pre-1511 Windows 10 without the newer API.
-    3. ``platform.machine()``.
-
-    Note ``GetNativeSystemInfo`` is deliberately NOT used: Microsoft documents
-    that it "also returns emulated processor details when run from an app
-    under emulation", so on the very WoA hosts this function exists to serve
-    it reports AMD64 — no better than the env-var rung below it.
-    """
-    if sys.platform == "win32":
-        try:
-            name = _windows_native_machine_from_iswow64()
-        except (OSError, AttributeError, TypeError, ValueError):
-            # API missing (pre-1511), DLL load failure in tests, or a
-            # mistyped ctypes binding — fall through to the env vars.
-            name = None
-        if name:
-            return name
-        env_arch = os.environ.get("PROCESSOR_ARCHITEW6432") or os.environ.get(
-            "PROCESSOR_ARCHITECTURE"
-        )
-        if env_arch:
-            return env_arch.upper()
-    import platform as _platform
-
-    return (_platform.machine() or "").upper()
-
-
-def _expected_windows_pe_machines() -> set:
-    """PE machine values the current Windows host can natively load.
-
-    Preferred source is ``GetMachineTypeAttributes``, which answers this
-    question directly (including AMD64-on-ARM64 emulation) instead of
-    inferring it from an architecture name.
-
-    Fallback is name-based: AMD64 hosts run x64 and (via WOW64) x86. ARM64
-    hosts run ARM64 and (Windows 11 emulation) x64. 32-bit x86 hosts run only
-    x86. Unknown machines return the permissive full set so the integrity gate
-    can never brick launch on exotic hosts. Host detection uses the OS-native
-    machine (see ``_windows_native_machine``), not the process architecture.
-    """
-    if sys.platform == "win32":
-        try:
-            runnable = _windows_user_runnable_pe_machines()
-        except (OSError, AttributeError, TypeError, ValueError):
-            runnable = None
-        if runnable:
-            return runnable
-    machine = _windows_native_machine().upper()
-    if machine in ("AMD64", "X86_64", "X64"):
-        return {_PE_MACHINE_AMD64, _PE_MACHINE_I386}
-    if machine in ("ARM64", "AARCH64"):
-        return {_PE_MACHINE_ARM64, _PE_MACHINE_AMD64}
-    if machine in ("X86", "I386", "I486", "I586", "I686"):
-        return {_PE_MACHINE_I386}
-    return {_PE_MACHINE_AMD64, _PE_MACHINE_ARM64, _PE_MACHINE_I386}
-
-
-def _parse_pe_machine(path: Path) -> int:
-    """Parse ``path`` as a PE executable and return its COFF machine field.
-
-    Raises ``ValueError`` with a human-readable reason when the file is not a
-    structurally complete PE: missing MZ/PE magic (an HTML error page or JSON
-    body saved as .exe), header truncation, or raw section data extending past
-    the end of the file (the truncated-download / interrupted-extraction
-    shape). Purely a header walk — cheap even on a 200 MB Electron exe.
-    """
-    import struct
-
-    try:
-        file_size = path.stat().st_size
-    except OSError as exc:
-        raise ValueError(f"unreadable: {exc}")
-    if file_size < 512:
-        raise ValueError(
-            f"file is only {file_size} bytes — far too small to be a Windows executable"
-        )
-    with path.open("rb") as fh:
-        head = fh.read(64)
-        if len(head) < 64 or head[:2] != b"MZ":
-            raise ValueError(
-                "missing MZ header — not a Windows executable "
-                "(a truncated or non-binary file saved as .exe?)"
-            )
-        e_lfanew = struct.unpack_from("<I", head, 0x3C)[0]
-        if e_lfanew <= 0 or e_lfanew + 24 > file_size:
-            raise ValueError("corrupt DOS header: PE header offset points past end of file")
-        fh.seek(e_lfanew)
-        pe_head = fh.read(24)
-        if len(pe_head) < 24 or pe_head[:4] != b"PE\x00\x00":
-            raise ValueError("missing PE signature — corrupt executable header")
-        machine, n_sections = struct.unpack_from("<HH", pe_head, 4)
-        size_of_optional = struct.unpack_from("<H", pe_head, 20)[0]
-        fh.seek(e_lfanew + 24 + size_of_optional)
-        max_section_end = 0
-        for _ in range(n_sections):
-            section = fh.read(40)
-            if len(section) < 40:
-                raise ValueError("truncated PE section table")
-            size_of_raw, pointer_to_raw = struct.unpack_from("<II", section, 16)
-            max_section_end = max(max_section_end, pointer_to_raw + size_of_raw)
-        if file_size < max_section_end:
-            raise ValueError(
-                f"truncated executable: file is {file_size} bytes but its PE "
-                f"sections extend to {max_section_end} bytes"
-            )
-    return machine
-
-
-def _pe_machine_or_none(path: Path) -> Optional[int]:
-    try:
-        return _parse_pe_machine(path)
-    except ValueError:
-        return None
-
-
-def _desktop_exe_integrity_error(path: Path) -> Optional[str]:
-    """Return a human-readable reason ``path`` cannot run on this Windows host,
-    or ``None`` when the exe parses as a complete PE of a loadable architecture.
-    """
-    try:
-        machine = _parse_pe_machine(path)
-    except ValueError as exc:
-        return str(exc)
-    expected = _expected_windows_pe_machines()
-    if machine not in expected:
-        got = _PE_MACHINE_NAMES.get(machine, f"unknown machine 0x{machine:04X}")
-        return (
-            f"architecture mismatch: built a {got} executable but this is a "
-            f"{_windows_native_machine()} Windows host"
-        )
-    return None
-
-
-def _desktop_backup_unpacked_dir(packaged_executable: Path) -> Path:
-    """The rollback tree before-pack.mjs preserves: ``<unpacked-dir>.bak``."""
-    unpacked = packaged_executable.parent
-    return unpacked.parent / (unpacked.name + ".bak")
-
-
-def _rollback_desktop_from_backup(packaged_executable: Path) -> Optional[Path]:
-    """Restore the previous unpacked desktop app from its ``.bak`` tree.
-
-    Returns the restored executable path, or ``None`` when no usable backup
-    exists (missing, or its exe fails the same integrity probe). The corrupt
-    tree is kept alongside as ``<unpacked-dir>.corrupt`` for diagnostics.
-    Best-effort: never raises.
-    """
-    unpacked = packaged_executable.parent
-    backup_dir = _desktop_backup_unpacked_dir(packaged_executable)
-    backup_exe = backup_dir / packaged_executable.name
-    if not backup_exe.exists():
-        return None
-    if _desktop_exe_integrity_error(backup_exe) is not None:
-        return None
-    corrupt_dir = unpacked.parent / (unpacked.name + ".corrupt")
-    try:
-        shutil.rmtree(corrupt_dir, ignore_errors=True)
-        try:
-            unpacked.rename(corrupt_dir)
-        except OSError:
-            shutil.rmtree(unpacked, ignore_errors=True)
-        backup_dir.rename(unpacked)
-    except OSError:
-        return None
-    restored = unpacked / packaged_executable.name
-    return restored if restored.exists() else None
-
-
-def _ensure_desktop_exe_launchable(
-    desktop_dir: Path, packaged_executable: Optional[Path]
-) -> tuple:
-    """Windows post-build integrity gate for the self-update rebuild (#69179).
-
-    Returns ``(verified_exe_or_None, rolled_back)``:
-
-    - exe passed the probe → ``(exe, False)``
-    - exe corrupt/wrong-arch, previous build restored → ``(old_exe, True)``
-    - exe corrupt and nothing restorable → ``(None, False)``
-
-    On any integrity failure the corrupt cached Electron zip is purged and the
-    desktop build stamp invalidated, so the updater's retry-once rebuild pulls
-    a fresh, SHASUM-verified Electron download instead of re-staging the same
-    corrupt bytes. No-op off Windows and when there is no executable to check.
-    """
-    if packaged_executable is None or sys.platform != "win32":
-        return packaged_executable, False
-
-    error = _desktop_exe_integrity_error(packaged_executable)
-    if error is None:
-        return packaged_executable, False
-
-    print(f"✗ The built Hermes.exe failed its integrity check: {error}")
-    print(f"    at: {packaged_executable}")
-
-    # Self-heal setup for the retry: drop the (likely corrupt) cached Electron
-    # zip and the content stamp so the next rebuild is a genuine re-download +
-    # re-stage rather than a replay of the same broken extraction.
-    _purge_electron_build_cache(desktop_dir)
-    try:
-        _desktop_stamp_path().unlink()
-    except OSError:
-        pass
-
-    restored = _rollback_desktop_from_backup(packaged_executable)
-    if restored is not None:
-        print("  ↩ Update aborted — restored the previous working Hermes.exe from backup.")
-        print("    Your existing version was kept and still works. Run `hermes desktop`")
-        print("    (or the in-app update) again to retry with a fresh Electron download.")
-        return restored, True
-
-    print("  ✗ No usable backup was found to restore.")
-    print("    Run `hermes desktop --force-build` to rebuild, or re-run the Hermes")
-    print("    installer to repair the install.")
-    return None, False
 
 
 def _electron_download_cache_dirs() -> list[Path]:
@@ -12566,6 +11768,39 @@ def main():
     else:
         parser.print_help()
 
+
+# ─── Web UI build + Windows exe-integrity clusters extracted to
+# hermes_cli/web_ui_build.py and hermes_cli/pe_integrity_mixin.py (main.py
+# decomposition, wave 1). Every moved name is re-exported here so callers and
+# existing test monkeypatches (hermes_cli.main._web_ui_build_needed,
+# hermes_cli.main._ensure_desktop_exe_launchable, ...) keep resolving
+# unchanged.
+from hermes_cli.web_ui_build import (  # noqa: E402
+    _build_web_ui,
+    _compute_desktop_content_hash,
+    _compute_web_ui_content_hash,
+    _desktop_build_needed,
+    _desktop_dist_exists,
+    _desktop_stamp_path,
+    _do_build_web_ui,
+    _missing_web_build_tool,
+    _web_ui_build_needed,
+    _web_ui_stamp_path,
+    _write_desktop_build_stamp,
+    _write_web_ui_build_stamp,
+)
+from hermes_cli.pe_integrity_mixin import (  # noqa: E402
+    _desktop_backup_unpacked_dir,
+    _desktop_exe_integrity_error,
+    _ensure_desktop_exe_launchable,
+    _expected_windows_pe_machines,
+    _parse_pe_machine,
+    _pe_machine_or_none,
+    _rollback_desktop_from_backup,
+    _windows_native_machine,
+    _windows_native_machine_from_iswow64,
+    _windows_user_runnable_pe_machines,
+)
 
 if __name__ == "__main__":
     main()
